% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-scale.R
\name{gm_mean}
\alias{gm_mean}
\title{Estimate Geometric Means}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{should NA be removed?

source:
https://stackoverflow.com/questions/2602583/geometric-mean-is-there-a-built-in}
}
\description{
for normalizers
}
