% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_fluidigm}
\alias{read_fluidigm}
\title{Parse the CSV Output of Fluidigm}
\usage{
read_fluidigm(path, simplify = TRUE)
}
\arguments{
\item{path}{The path to the CSV file}

\item{simplify}{TRUE or FALSE.

  If TRUE, it returns a simplified version of the dataset
  that contains only essential variables.}
}
\description{
Parse the output of the Fluidigm qPCR measurement
from CSV format into R.
This function uses `read_delim()` from the package
`readr` to read the data
and functions from the `dplyr` package to perform some
basic tidying
}
