% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-scale.R
\name{scale_01_fluidigm}
\alias{scale_01_fluidigm}
\title{Scale Normalized Fluidigm data}
\usage{
scale_01_fluidigm(.data, .group)
}
\arguments{
\item{.data}{A `data.frame` produced by `read_fluidigm()` and normalized by
`normalize()`. The columns required
are: `sample_name`,`target_name` and `expression`.}

\item{.group}{A column of `.data` that you want to use to group your
expression values before scaling.}
}
\description{
Scale your fluidigm data in range 0 to 1.
Run this function after `normalize()`.
}
