% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-scale.R
\name{scale_ddct}
\alias{scale_ddct}
\title{Center Expression on One Single Experimental Condition}
\usage{
scale_ddct(.data, group_var, group_var2, compare_var, center_on)
}
\arguments{
\item{.data}{A `data.frame` produced by `read_fluidigm()` and normalized by
`normalize()`. The columns required
are: `sample_name`,`target_name` and `expression`.}

\item{compare_var}{the column of the dataset that stores the variable that
you want to use for centering}

\item{center_on}{the value of `compare_var` that you want to set to 1 (i.e.
the one that you want to center your data on)}
}
\description{
Add the `scaled_ddct` columns which stores delta delta Ct
scaled values
}
