% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-scale.R
\name{normalize}
\alias{normalize}
\title{Normalize Ct Values}
\usage{
normalize(.data, normalizers)
}
\description{
Normalize Ct values in each sample on the geometric mean of user provided
reference genes.
}
