% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-scale.R
\name{scale_ddct}
\alias{scale_ddct}
\title{Center Expression on One Single Experimental Condition}
\usage{
scale_ddct(.data, group_var, group_var2, center_var)
}
\arguments{
\item{.data}{A `data.frame` produced by `read_fluidigm()` and normalized by
`normalize()`. The columns required
are: `sample_name`,`target_name` and `expression`.}
}
\description{
Add the `scaled_ddct` columns which stores delta delta Ct
scaled values
}
